+++ breeze-reflect/src/main/kotlin/com/windea/breezeframework/reflect/extensions/ReflectExtensions.kt
@@ -19,13 +19,13 @@ fun checkClassForName(className: String): Boolean {
 
 
 /**得到指定类型的名字。*/
-@TrickImplementationApi("Can never be implemented for all situations.")
+@TrickImplementationApi("Cannot get actual name of a function parameter or a local variable.")
 inline fun <reified T> nameOf(): String? {
 	return T::class.java.simpleName
 }
 
 /**得到指定项的名字。适用于：类引用、属性引用、方法引用、实例。不适用于：类型参数，参数，局部变量。*/
-@TrickImplementationApi("Can never be implemented for all situations.")
+@TrickImplementationApi("Cannot get actual name of a function parameter or a local variable.")
 @JvmSynthetic
 inline fun nameOf(target: Any?): String? {
 	//无法直接通过方法的引用得到参数，无法得到局部变量的任何信息